Renamed dto_model to data_transfer_object_model, along with it's relevant class. Added comment to describe literal passed to annotation in Http SessionConfig (cannot pass constants to annotations.) Removed /api from all controller mappings.
