+++ server/src/main/java/com/creppyfm/server/controller/TaskController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/tasks")
+@RequestMapping("/tasks")
 @Tag(name = "Task Controller")
 public class TaskController {
