+++ server/src/main/java/com/creppyfm/server/data_transfer_object_model/UserDataTransferObject.java
@@ -1,4 +1,4 @@
-package com.creppyfm.server.dto_model;
+package com.creppyfm.server.data_transfer_object_model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class UserDTO {
+public class UserDataTransferObject {
     private String firstName;
     private String lastName;
     private String email;
